chore(gltf): gltf module cleanup
