+++ examples/tutorials/hello-gltf/index.html
@@ -33,6 +33,12 @@
     #model-options label {
       color: white;
     }
+    #error {
+      position: absolute;
+      left: 10px;
+      top: 120px;
+      color: red;
+    }
   </style>
 </head>
 <body>
@@ -43,4 +49,5 @@
     <div><label><input type="checkbox" id="gltfAnimation" />glTF Animation</label></div>
     <div><label><input type="checkbox" id="pbr" />PBR</label></div>
   </div>
+  <div id="error"></div>
 </body>
