+++ modules/gltf/src/index.ts
@@ -1,10 +1,10 @@
 // luma.gl, MIT license
 
-export type {PBREnvironment} from './pbr/pbr-environment';
-export type {ParsePBRMaterialOptions, ParsedPBRMaterial} from './pbr/parse-pbr-material';
-export {parsePBRMaterial} from './pbr/parse-pbr-material';
-export {loadPBREnvironment} from './pbr/pbr-environment';
+export {loadPBREnvironment, type PBREnvironment} from './pbr/pbr-environment';
+export {type ParsedPBRMaterial} from './pbr/pbr-material';
+export {parsePBRMaterial, type ParsePBRMaterialOptions} from './parsers/parse-pbr-material';
+export {} from './pbr/pbr-environment';
 
 // glTF Scenegraph Instantiator
-export {createScenegraphsFromGLTF} from './gltf/create-gltf-objects';
+export {createScenegraphsFromGLTF} from './gltf/create-scenegraph-from-gltf';
 export {GLTFAnimator} from './gltf/gltf-animator';
