+++ package.json
@@ -64,5 +64,6 @@
   "volta": {
     "node": "20.17.0",
     "yarn": "4.9.1"
-  }
+  },
+  "packageManager": "yarn@4.9.1"
 }
